+++ type-dossier/update_status_pes_s2low/TypeDossierUpdateStatusPesS2lowEtape.php
@@ -12,9 +12,13 @@ final class TypeDossierUpdateStatusPesS2LowEtape implements TypeDossierEtapeSetS
         StringMapper $stringMapper,
     ): array {
         $changeStatusAction = $stringMapper->get('change-status-transaction');
+        $changeStatutErrorState = $stringMapper->get('change-status-error');
 
         $result[DocumentType::ACTION][$changeStatusAction]
         [Action::CONNECTEUR_TYPE_MAPPING]['transaction'] = $typeDossierEtape->specific_type_info['transaction'];
+
+        $result[DocumentType::ACTION]['supression'][Action::ACTION_RULE]
+        [Action::ACTION_RULE_LAST_ACTION][] = $changeStatutErrorState;
 
         return $result;
     }