+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Corrections
 
 - Le connecteur de récupération des arriérés actes de s2low ne fonctionnait pas avec un login/mdp #2221
+- Modification des libellés des dates du connecteur de récupération des arriérés actes de s2low #2217
 - Il était possible de créer un connecteur d'entité au niveau de l'entité racine même si la fonctionnalité était désactivée #2224
 
 # [4.1.6] - 2025-01-28